Move perform into sync.go
