+++ pkg/skaffold/sync/kubectl/kubectl.go
@@ -21,28 +21,25 @@ import (
 	"fmt"
 	"os/exec"
 
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/sync"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
 
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"k8s.io/api/core/v1"
-	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-type KubectlSyncer struct{}
+type Syncer struct{}
 
-func (k *KubectlSyncer) Sync(ctx context.Context, s *sync.Item) error {
+func (k *Syncer) Sync(ctx context.Context, s *sync.Item) error {
 	logrus.Infoln("Copying files:", s.Copy, "to", s.Image)
 
-	if err := perform(ctx, s.Image, s.Copy, copyFileFn); err != nil {
+	if err := sync.Perform(ctx, s.Image, s.Copy, copyFileFn); err != nil {
 		return errors.Wrap(err, "copying files")
 	}
 
 	logrus.Infoln("Deleting files:", s.Delete, "from", s.Image)
 
-	if err := perform(ctx, s.Image, s.Delete, deleteFileFn); err != nil {
+	if err := sync.Perform(ctx, s.Image, s.Delete, deleteFileFn); err != nil {
 		return errors.Wrap(err, "deleting files")
 	}
 
@@ -55,45 +52,4 @@ func deleteFileFn(ctx context.Context, pod v1.Pod, container v1.Container, src, 
 
 func copyFileFn(ctx context.Context, pod v1.Pod, container v1.Container, src, dst string) *exec.Cmd {
 	return exec.CommandContext(ctx, "kubectl", "cp", src, fmt.Sprintf("%s/%s:%s", pod.Namespace, pod.Name, dst), "-c", container.Name)
-}
-
-func perform(ctx context.Context, image string, files map[string]string, cmdFn func(context.Context, v1.Pod, v1.Container, string, string) *exec.Cmd) error {
-	if len(files) == 0 {
-		return nil
-	}
-
-	client, err := kubernetes.Client()
-	if err != nil {
-		return errors.Wrap(err, "getting k8s client")
-	}
-
-	pods, err := client.CoreV1().Pods("").List(meta_v1.ListOptions{})
-	if err != nil {
-		return errors.Wrap(err, "getting pods")
-	}
-
-	synced := map[string]bool{}
-
-	for _, p := range pods.Items {
-		for _, c := range p.Spec.Containers {
-			if c.Image != image {
-				continue
-			}
-
-			for src, dst := range files {
-				cmd := cmdFn(ctx, p, c, src, dst)
-				if err := util.RunCmd(cmd); err != nil {
-					return err
-				}
-
-				synced[src] = true
-			}
-		}
-	}
-
-	if len(synced) != len(files) {
-		return errors.New("couldn't sync all the files")
-	}
-
-	return nil
 }
